converted validation function to custom hook
