+++ src/components/form/DateInput.tsx
@@ -20,6 +20,7 @@ const DateInput: React.FC<DateProps> = ({ id, label }) => {
     <motion.div
       variants={{ hidden: { opacity: 0, scale: 0.5 }, visible: { opacity: 1, scale: 1 } }}
       transition={{ type: 'easeIn', duration: 0.5 }}
+      whileHover={{ scale: 1.1 }}
     >
       <DatePicker
         className={css.input}
