+++ src/components/form/Form.module.css
@@ -29,6 +29,6 @@
   background-color: transparent;
   text-shadow: 1px 2px 2px rgb(23, 23, 23);
   border-radius: 10px;
-  padding: 0.5rem;
+  padding: 0;
   border: none;
 }
